moved tracker components into folder
