+++ moby_app/src/app/tracker/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import AuditionHistory from '@/components/AuditionHistory';
+import AuditionHistory from '@/components/tracker/AuditionHistory';
 
 export default function TrackerPage() {
     return (
